improve range checking

+++ runtime/devicescript/buffer.c
@@ -80,7 +80,7 @@ static double clamp_double(value_t v, double l, double h) {
         F64 = SZ;                                                                                  \
         break;
 
-value_t jacs_buffer_op(jacs_activation_t *frame, uint16_t fmt0, uint16_t offset, uint16_t buffer,
+value_t jacs_buffer_op(jacs_activation_t *frame, uint32_t fmt0, uint32_t offset, uint32_t buffer,
                        value_t *setv) {
     int is_float = 0;
